RISC-V: Networking fix Hack
It looks like that kernel driver now supports reseting the
signal one additional time. As it had been  already reset
twice in FSBL, PHY gets into incorrect state causing below error.

----------------------------------------------------------------------
macb 10090000.ethernet (unnamed net_device) (uninitialized): Could not attach to PHY
macb: probe of 10090000.ethernet failed with error -110
----------------------------------------------------------------------

This patch is just a temporary fix until we have a fix a FSBL.
It is just a **HACK** and **NOT TO BE MERGED** into mainline.

Signed-off-by: Atish Patra <atish.patra@wdc.com>